moves shredder out of the shred module
Working towards embedding versioning into shreds binary; so Shredder can
no longer depend on internals of Shred.
