+++ ledger/src/lib.rs
@@ -24,6 +24,7 @@ pub mod next_slots_iterator;
 pub mod rooted_slot_iterator;
 pub mod shred;
 pub mod shred_stats;
+mod shredder;
 pub mod sigverify_shreds;
 pub mod slot_stats;
 pub mod staking_utils;
